moved RO_with_energy_recovery examples
